server/biomarker_requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
PyPDF2>=3.0.0
pdfplumber>=0.11.0
